requirements.txt
# OpenSearch client
opensearch-py
# Fake data generator
faker
# Dotenv for environment variables
python-dotenv
